Change deploy script to rename the executable before service restart
Because I want to minimize down time, now we only have to wait for
the service restart. You can rename a file while it is executed,
the file is still accessible for the process executing the file,
the HDD reference is the same, it just has a different name.

The reason for keeping a old version is to be able to quickly revert
if a deploy fails.